requirements.txt
altgraph==0.17
certifi==2020.6.20
chardet==3.0.4
future==0.18.2
idna==2.10
lz4==3.1.0
pefile==2019.4.18
Pillow==7.2.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
PySide2==5.15.1
pywin32-ctypes==0.2.0
requests==2.24.0
shiboken2==5.15.1
urllib3==1.25.10
